chore: extract ActionModeClass from NoteEditor
